==> 04b_TrainingPt_Summary.Rmd <==
---
title: "Field transect data summaries"
subtitle: "by Gen Perkins"
date: "10/21/2020"
output: html_document
---

```{r global_options, include=FALSE , warnings = FALSE, messgaes = FALSE}
require(knitr)
knitr::opts_chunk$set(message = FALSE)

```


```{r setup, include=FALSE}

library(tidyverse)
library(raster)
library(sf)
library(stringr)
library(lwgeom)

```

```{r set up folder structure, include = FALSE}
## 1. Set up folder structure for Stage 1 sampling 

AOI <- "KootInvCran"
#AOI <- "Deception"
#AOI <- "EagleHills"
#AOI <- "DateCreek"


AOI_dir <- file.path(".", paste0(AOI,"_AOI"))

clean_pts <- file.path(AOI_dir, "1_map_inputs","trainingData","clean")

s1_dir <- file.path(AOI_dir, "2_sample_design", "stage1_StudyDesign", "training_pts")

#s2_dir <- file.path(AOI_dir, "2_sample_design", "stage2_StudyDesign", "training_pts")
  
points_dir <- file.path(AOI_dir, "2_sample_design", "stage2_point_data", "training_pts")

map.key  <- read.csv(file.path(AOI_dir, "_MapUnitLegend", 
                                 paste0(AOI, "_MapUnitLegend.csv")), 
                       stringsAsFactor = FALSE)
```

### Training data summary report : `r AOI`.

The following report summarises field calls by length (transects). A summary of raw point calls, on which transects were collected is include in the Appendices. 

```{r import base files to run analysis, tidy = TRUE, warning=FALSE, include = FALSE}
## 1) Import consolidated data and linear transects

s1_trans <- list.files(s1_dir, pattern = "s1_transects", full.names = TRUE, recursive = FALSE)[16]
s1_pts <- list.files(s1_dir, pattern = "s1_pts.gpkg$", full.names = TRUE, recursive = FALSE)[16]

#s2_trans <- list.files(s2_dir, pattern = "^proc", full.names = TRUE, recursive = FALSE)
#s2_pts <- list.files(s2_dir, pattern = "_pts.gpkg$", full.names = TRUE, recursive = FALSE)

```


### 1) Summary of line data length

```{r, echo = FALSE, messgae = FALSE}
# Deception: 
all_trans <- st_read(s1_trans, quiet = TRUE)
#s2_trans <- st_read(s2_trans, quiet = TRUE)
#all_trans <- bind_rows(s1_trans, s2_trans) 


all_trans <- all_trans %>%  
    mutate(length = st_length(.)) %>%
    st_drop_geometry() 


#all_trans_i <- all_trans %>%
#  dplyr::filter(data_type == "incidental")

# total type of points with primary and secondary 
trans_type_p <- all_trans %>%
  dplyr::group_by(data_type) %>%
  dplyr::summarise(primary = sum(length, na.rm = TRUE))

trans_type_s <- all_trans %>%
  filter(!is.na(mapunit2)) %>%
  group_by(data_type) %>%
  dplyr::summarise(secondary = sum(length, na.rm = TRUE))

trans_type <- left_join(trans_type_p, trans_type_s) %>%
  mutate(pc.sec = round((secondary/primary * 100),1))

trans_type_long <- left_join(trans_type_p, trans_type_s) %>%
  mutate(pc.sec = secondary/primary * 100)


trans_types <- unique(trans_type$data_type)

```
 
A total length of `r round(sum(trans_type$primary),1)` meters was surveyed during transect collection. This include `r round(length(trans_type$data_type),1)` data types : `r trans_types`.

Alternate calls were made for `r round(sum(trans_type$secondary, na.rm = TRUE),2)` meters, with an average of `r round(sum(trans_type$secondary, na.rm = TRUE)/sum(trans_type$primary, na.rm = TRUE)*100, 1) ` % of all calls.

```{r, results  = "asis", hide = TRUE, echo = FALSE}

kable(trans_type)

```

```{r, echo = FALSE, fig.width = 10, fig.height=10}
#all_trans

trans_f<- all_trans %>%
  mutate(mapunit2 = ifelse(mapunit2 == "", NA, mapunit2),
         mapunit1 = ifelse(mapunit1 == "", NA, mapunit1)) %>%
  group_by(mapunit1, mapunit2) %>%
  dplyr::summarise(length = sum(length)) %>%
  ungroup()
  
# primary calls: 
trans_f_p <- trans_f %>%
  filter(is.na(mapunit2)) 

# format secondary calls:
trans_f_s12 <- trans_f %>%
  filter(!is.na(mapunit2)) %>%
  mutate(order = "mapunit12")
       
trans_f_s21 <- trans_f %>%
  filter(!is.na(mapunit2)) %>%
  mutate(order = "mapunit21",
         mapunit3 = mapunit1, 
         mapunit4 = mapunit2) %>%
  dplyr::select(-c(mapunit1, mapunit2)) %>%
  dplyr::rename(mapunit1 = mapunit4, 
                mapunit2 = mapunit3)

trans_secondary <- bind_rows(trans_f_s12, trans_f_s21) %>%
  arrange(mapunit1, mapunit2) %>%
  group_by(mapunit1, mapunit2) %>%
  dplyr::summarise(length = sum(length)) 

trans_secondary_f1 <- trans_secondary %>%
  group_by(length) %>%
  arrange(length, group_by = TRUE) %>%
  aggregate(.~length, . , FUN=head, 1) %>%
  dplyr::select(mapunit1, mapunit2, length)%>%
  arrange(mapunit1) %>%
  mutate(length = round(length, 2))


trans_summary <- bind_rows(trans_f_p, trans_secondary_f1) 
  
trans_total <- trans_summary %>%
  arrange(mapunit1) %>%
  rowwise() %>%
  mutate(bgc = unlist(strsplit(mapunit1, "_"))[1]) %>%
  mutate(bgc = ifelse(str_detect(mapunit1, "_"), bgc, "non-forest")) %>%
  ungroup() %>%
  mutate(length = as.numeric(round(length, 1))) %>%
  filter(!is.na(mapunit1)) %>%
  mutate(mapunit12 = ifelse(is.na(mapunit2), mapunit1, paste0(mapunit1, "/", mapunit2)))

bgc_list <- unique(trans_total$bgc)

```

The length of calls per bgc varied, along with proportion of primary and secondary calls. Figures below show the length of calls for all data types grouped by predominant bgc.


```{r, echo = FALSE, fig.width= 12, fig.height = 16}
# plot the freqency 

# points collected per type (primary and secondary calls) all calls
#p <- ggplot(trans_total, aes(y = mapunit12, x = length)) + 
#  facet_wrap(~bgc, scales = "free_x") +    
#  geom_bar(stat = "identity")
#         
#print(p)



for(b in bgc_list) {

 # b = unique(pt_total$bgc)[3]
  pt_total_bgc <- trans_total %>%
    filter(bgc == b)
  
  p <- ggplot(pt_total_bgc , aes(y = reorder(mapunit12, length), x = length)) +
  geom_bar(stat = "identity") + 
  labs(title = b, y = "mapunit")
  
  print(p)
  
#  ggplot(pt_total_bgc , aes(y = mapunit1, x = freq)) +
#  geom_bar(stat = "identity")
     
  #ggsave(file.path(s1_dir, paste0(b, "_point_freq_s1.png")))
  
  
}

```


### Proportion of primary and secondary calls per observer.

```{r echo = FALSE}

# group by observer

all_trans_ob <- all_trans %>%
  group_by(transect_id) %>%
  fill( observer, .direction = "downup")

all_trans_ob_sum <- all_trans_ob %>%
   group_by(transect_id, observer) %>%
    dplyr::summarise(count = n())

trans_ob_p <- all_trans_ob %>%
 # dplyr::filter(is.na(mapunit2)) %>%
  group_by(observer) %>%
  summarise(all.calls = sum(length))

trans_ob_s <- all_trans_ob %>%
  dplyr::filter(!is.na(mapunit2)) %>%
  group_by(observer) %>%
  summarise(sec.calls = sum(length))


trans_obs_type <- left_join(trans_ob_p, trans_ob_s) %>%
  mutate(pc.sec = round((sec.calls/all.calls * 100),1))

trans_obs_type<- trans_obs_type %>%
  mutate(observer = case_when(
    observer == "EAC GP" ~ "EAC",
    observer == "EAC GCP" ~ "EAC",
    observer == "EC" ~ "EAC",
    observer == "CC LJ" ~ "CC",
    observer == "DF LJ" ~ "DF",
    observer =="DJM DF" ~ "DJM",
    observer == "PD AR" ~ "PD", 
    observer =="SRan" ~ "SR",
    observer == "SR/MZ" ~ "SR",
    observer == "WM" ~ "WHM",
        TRUE ~ observer)) %>%
      group_by(observer) %>%
      summarise(across(where(is.numeric), sum)) %>%
  mutate(pc.sec = round(sec.calls/all.calls * 100,1))

```


There were `r length(trans_ob_s$observer)` observers recording data, including unnamed recorders (NA). The table below shows the total calls and the proportion of calls which also had a secondary call. The average proportion of transect length that has a secondary call was `r mean(trans_ob_s$pc.sec)`. Note this data does not include repeat transects, in which multiple observers recorded at the same location. 

```{r, results  = "asis", hide = TRUE, echo = FALSE}

kable(trans_obs_type)

```




```{r, include = FALSE}
# 
# Appendices
# ### Summary of point data 
# 
# Point data represents the raw data collected by observers during transect traverses. This is included for comparison only. 

```


```{r, include = FALSE}
# combine stage 1 and 2 and summarise calls

s1_pt <- st_read(s1_pts, quiet = T) 
#s2_pt <- st_read(s2_pts, quiet = T) 

#pts <- bind_rows(s1_pt, s2_pt) %>%
pts <- s1_pt %>%
  dplyr::select(-order) %>%
  st_drop_geometry() %>%
  filter(!is.na(mapunit1)) %>%
  mutate(mapunit12 = paste0(mapunit1,"_", mapunit2)) %>%
  mutate(mapunit12 = gsub("_NA","", mapunit12)) %>%
  mutate(mapunit12 = gsub("_$","", mapunit12)) 

# proportion of calls with primary and secondary calls

# total type of points with primary and secondary 
pt_type_p <- pts %>%
  group_by(data_type) %>%
  dplyr::summarise(all.calls = n())

#pt_type_i <- pts %>%
#  filter(data_type == "incidental")

pt_type_s <- pts %>%
  filter(!is.na(mapunit2)) %>%
  group_by(data_type) %>%
  dplyr::summarise(sec.calls = n())

pt_type <- left_join(pt_type_p, pt_type_s) %>%
  mutate(percent.sec = sec.calls/all.calls * 100)

pt_types <- unique(pt_type$data_type)

```

```{r, include = FALSE}

# At `r AOI` study area a total `r sum(pt_type$all.calls)` point calls were collected during traverse (stage 1 and stage 2) transect collection. This include `r length(pt_type$data_type)` data types : `r pt_types`.
# 
# Of these, `r sum(pt_type$sec.calls)` points were also given an secondary call, representing an average percent of `r round(sum(pt_type$sec.calls)/sum(pt_type$all.calls)*100, 1) ` %).

```


```{r, results  = "asis", hide = TRUE, echo = FALSE,  include = FALSE}

kable(pt_type)

```

```{r, echo = FALSE, warnings = FALSE, messages = FALSE, fig.width = 12,fig.height = 16, include = FALSE}
# calculate the frequency of all calls with primary and secondary
#The following figures provide a breakdown of primary and secondary call #frequency of point data collected. Note in the figures below secondary #calls are presented twice in each figure (ie: CX_AM and AM_CX). 

# stage 1 
pt_f <- pts %>%
  filter(data_type == "s1") %>%
  mutate(mapunit2 = ifelse(mapunit2 == "", NA, mapunit2),
         mapunit1 = ifelse(mapunit1 == "", NA, mapunit1)) %>%
  group_by(mapunit1, mapunit2) %>%
  dplyr::summarise(freq = n())
  

# primary calls: 
pt_f_p <- pt_f %>%
  ungroup() %>%
  filter(is.na(mapunit2)) %>%
  dplyr::select(-mapunit2)

# format secondary calls:
pt_f_s <- pt_f %>%
  filter(!is.na(mapunit2)) %>%
  mutate(mapunit12 = paste0(mapunit1, "_", mapunit2),
         mapunit21 = paste0(mapunit2, "_", mapunit1)) %>%
  ungroup()

# grab mapunit12
pt_f_s1 <- pt_f_s %>%
  dplyr::select(mapunit12, freq) %>%
  mutate(type = "mapunit12")

# grab mapunit21
pt_f_s2 <- pt_f_s %>%
  dplyr::select(mapunit21, freq)%>%
  mutate(type = "mapunit21",
         mapunit12 = mapunit21 ) %>%
 dplyr::select(-c(mapunit21))

#duplicates <- c(pt_f_s2$mapunit12)

pt_secondary <- bind_rows(pt_f_s1, pt_f_s2) %>%
  arrange(mapunit12) %>%
  group_by(mapunit12) %>%
  dplyr::summarise(freq = sum(freq)) %>%
  dplyr::rename(mapunit1 = mapunit12)

#pt_secondary_tidy <- pt_secondary %>%
#  mutate(map1 = str_split(mapunit1, "[:digit:]\\_", n = 2))
#pt_secondary_tidy$map1


#note these appear twice for each combination! Need to remove these!

pt_total <- bind_rows(pt_f_p, pt_secondary) %>%
  arrange(mapunit1) %>%
mutate(bgc = case_when(
    str_detect(mapunit1,"ESSFmc_") ~ "ESSFmc",
    str_detect(mapunit1, "SBS") ~  "SBS",
    str_detect(mapunit1,"ESSFdc2") ~ "ESSFdc",
    str_detect(mapunit1,"ESSFmh") ~ "ESSFmh",
    str_detect(mapunit1,"ICHmk1") ~ "ICHmk",
    str_detect(mapunit1,"IDFdm1") ~ "IDFdm",
    str_detect(mapunit1,"MSdm1") ~ "MSdm",
    TRUE ~ "non-forest"
    )) %>%
  dplyr::filter(!is.na(mapunit1))
  
# points collected per type (primary and secondary calls) all calls
#p <- ggplot(pt_total, aes(y = mapunit1, x = freq)) + 
#  facet_wrap(~bgc, scales = "free_x") +    
#  geom_bar(stat = "identity")
#         
#p

# points collected per bgc (primary and secondary calls) 
bgc_list <- unique(pt_total$bgc)
plot_list <- vector(length = length(bgc_list), mode = "list")
names(plot_list) <- bgc_list


for(b in bgc_list) {

 # b = unique(pt_total$bgc)[3]
  pt_total_bgc <- pt_total %>%
    filter(bgc == b) %>%
    arrange(freq)
  
  p <- ggplot(pt_total_bgc , aes(y = reorder(mapunit1, freq), x = freq)) +
  geom_bar(stat = "identity") + 
  labs(title = b, y = "mapunit")
  
  print(p)
  
#  ggplot(pt_total_bgc , aes(y = mapunit1, x = freq)) +
#  geom_bar(stat = "identity")
     
  #ggsave(file.path(s1_dir, paste0(b, "_point_freq_s1.png")))
  
}

```


### Summary of number of transects per BGC per slice: 

```{r, include=FALSE}
# 
# s1_pt 
# 
# unique(s1_pt$transect_id)
# 
# trans <- s1_pt %>% 
#   st_drop_geometry() %>%
#   group_by(transect_id)%>% 
#   summarise(sum = n()) %>%
#   filter(!is.na(transect_id))
#   
# #aa <- unique(s1_pt$transect_id)
# #str_split(aa, "_")
# 
# aa <- trans %>% 
#   mutate(transect_id = gsub("ICH mc", "ICHmc", transect_id)) %>%
#   mutate(temp = str_split(transect_id, "_")) %>%
#   rowwise() %>%
#   mutate(temp1 = temp[[1]]) %>%
#   mutate(temp2 = temp[[2]]) %>%
#   mutate(temp3 = temp[[3]]) 
# 
# aa_essf <- aa %>%
#   filter(str_detect(transect_id, "ESSF")) %>%
#   mutate(rotation = temp3, 
#          total = temp2) %>%
#   dplyr::select(-c(temp2,temp3))
# 
# aa_essf <- aa_essf %>%
#   mutate(temp2 = str_split(temp1, " ")) %>%
#   mutate(temp3 = temp2[[2]]) %>%
#   mutate(bgc = temp2[[1]])%>%
#   dplyr::select(-c(temp,temp2)) %>%
#   mutate(temp4 = str_split(temp3, "\\.")) %>%
#   mutate(slice = temp4[[1]]) %>%
#   mutate(site = temp4[[2]]) %>%
#   dplyr::select(c(transect_id, bgc, sum, rotation, slice, site, total))
#   
#   
# aa_ich <- aa %>%
#   filter(str_detect(transect_id, "ICH")) %>%
#   mutate(rotation = temp[[4]]) %>%
#   mutate(total = temp3) %>%
#   mutate(bgc = temp[[1]])%>%
#   dplyr::select(-c(temp, temp1,temp3)) %>%
#   mutate(temp4 = str_split(temp2, "\\.")) %>%
#   mutate(slice = temp4[[1]]) %>%
#   mutate(site = temp4[[2]]) %>%
#   dplyr::select(c(transect_id, bgc, sum, rotation, slice, site, total))
# 
# all <- rbind(aa_ich, aa_essf)
# 
# 
# sum <- all %>%
#   dplyr::select(bgc, slice, site, total)%>%
#   group_by(bgc)%>% 
#   summarise(slices = length(unique(slice)),
#             sites = length(unique(site)),
#             total = length(unique(total))
#   
# 

# # set up slices and sites per site 
# s1_pt <- s1_pt %>% 
#   mutate()
# 
# 
# trDat_slices <- trDat %>%
#   group_by (slice) %>%
#   dplyr::summarise(n.transect = length(unique(transect_id)),
#                    n.sites = length(unique(tid))) %>% 
#   pivot_longer(cols = c("n.transect", "n.sites"), names_to = "type", values_to = "number")
# 
# ggplot(trDat_slices, aes(slice, number, fill = type)) +
#   geom_bar(stat = "identity", position = "dodge") + 
#   theme(axis.text.x = element_text(angle = 90))+
#   theme_pem()+
#   scale_fill_discrete_sequential(palette = "Light Grays")



```

Speed per transect 

```{r}

library(maptools)
kml_dir <- "D:/PEM_DATA/BEC_DevExchange_Work/KootInvCran_AOI/2_sample_design/stage1_StudyDesign/transect_data/kmls/"

kml_out_dir <- file.path(kml_dir, "out")

kml<- list.files(path=kml_dir, pattern="*.kml", full.names=TRUE, recursive = TRUE)[1]

kmlsf <- sf::st_read(kml)


# To get altitude from KML, then mention ignoreAltitude=FALSE                

# delete the .kmz part from the column names in LonLat matrix
colnames(LonLat) <- gsub(pattern=".kmz", replacement="", x=colnames(LonLat))

# give names to rows
rownames(LonLat) <- c("Longitude", "Latitude")

```
